Replace inline version gen with js bar

+++ deploy.rb
@@ -5,6 +5,11 @@ if ARGV.length == 0
   exit(0)
 end
 
+if ENV['AWS_CLOUDFRONT_DIST_ID'] == '' || ENV['AWS_S3_BUCKET'] == ''
+  puts "AWS_CLOUDFRONT_DIST_ID and AWS_S3_BUCKET env vars are required"
+  exit(0)
+end
+
 `rm -rf build`
 ENV['RIAK_VERSION'] = ARGV[0]
 ENV['RIAKCS_VERSION'] = ARGV[1] || ARGV[0]
